work in progress

+++ week4/index.js
@@ -1,40 +1,15 @@
 const isAllCheckboxElement = document.querySelector('.todo__is-all__checkbox')
+const todoListElement = document.querySelector('.todo-list')
 const createTaskFormElement = document.forms.createTask
+// const allTasks = document.querySelectorAll('.todo-task')
 
-window.addEventListener('load', getTasks)
-
-document.addEventListener('change', (event) => {
-  const target = event.target
-
-  // set task done
-  if (target.classList.contains('todo-task-checkbox')) {
-    setTaskDone(target)
-  }
-
-  // show and hide done tasks
-  if (target === isAllCheckboxElement) {
-    showAllTasks()
-  }
+window.addEventListener('load', () => {
+  getTasks().then((tasks) => tasks.forEach(appendTaskToList))
 })
-
-document.addEventListener('click', (event) => {
-  const target = event.target
-
-  // remove task
-  if (target.classList.contains('todo-task__remove-btn')) {
-    removeTask(target)
-  }
-})
-
-document.addEventListener('submit', (event) => {
-  const target = event.target
-
-  event.preventDefault()
-
-  if (target === createTaskFormElement) {
-    createTask(target)
-  }
-})
+todoListElement.addEventListener('change', setTaskDone)
+isAllCheckboxElement.addEventListener('change', showAllTasks)
+todoListElement.addEventListener('click', removeTask)
+createTaskFormElement.addEventListener('submit', createTask)
 
 function appendTaskToList(task) {
   const taskTemplate = document
@@ -66,13 +41,8 @@ function appendTaskToList(task) {
   if (task.due_date) {
     const currentTime = new Date().getTime()
     const dueTime = new Date(task.due_date).getTime()
-    const dateString = new Date(task.due_date).toISOString().split('T')[0]
-    const timeString = new Date(task.due_date)
-      .toISOString()
-      .split('T')[1]
-      .split('.')[0]
 
-    dueDateElement.textContent = 'Due date: ' + dateString + ' ' + timeString
+    dueDateElement.textContent = 'Due date: ' + formatDate(task.due_date)
 
     if (currentTime > dueTime && !task.done)
       dueDateElement.classList.add('todo-task__due-date--overdue')
@@ -84,19 +54,23 @@ function appendTaskToList(task) {
   todoListElement.append(taskTemplate)
 }
 
-function setTaskDone(target) {
-  const taskElement = target.closest('.todo-task')
-  const id = taskElement.getAttribute('id').split('-')[1]
+function formatDate(date) {
+  return new Date(date).toLocaleDateString()
+}
 
-  taskElement.classList.toggle('todo-task--done')
+function setTaskDone(event) {
+  const target = event.target
 
-  if (taskElement.classList.contains('todo-task--done')) {
-    updateTask(id, true)
-  } else {
-    updateTask(id, false)
+  if (target.classList.contains('todo-task-checkbox')) {
+    const taskElement = target.closest('.todo-task')
+    const id = taskElement.getAttribute('id').split('-')[1]
+
+    taskElement.classList.toggle('todo-task--done')
+
+    updateTask(id, taskElement.classList.contains('todo-task--done'))
+
+    if (isAllCheckboxElement.checked) taskElement.style.display = 'block'
   }
-
-  if (isAllCheckboxElement.checked) taskElement.style.display = 'block'
 }
 
 function showAllTasks() {
@@ -109,7 +83,11 @@ function showAllTasks() {
   }
 }
 
-function createTask(target) {
+function createTask(event) {
+  event.preventDefault()
+
+  const target = event.target
+  // const id = allTasks.length + 1
   const title = target.title.value.trim()
   const done = false
   const description = target.description.value.trim() || undefined
@@ -120,6 +98,7 @@ function createTask(target) {
   if (!title) return
 
   const task = {
+    id,
     title,
     done,
     due_date,
@@ -132,42 +111,35 @@ function createTask(target) {
   target.reset()
 }
 
-function removeTask(target) {
-  const taskElement = target.closest('.todo-task')
-  const id = taskElement.getAttribute('id').split('-')[1]
+function removeTask(event) {
+  if (event.target.classList.contains('todo-task__remove-btn')) {
+    const taskElement = event.target.closest('.todo-task')
+    const id = taskElement.getAttribute('id').split('-')[1]
 
-  taskElement.remove()
+    taskElement.remove()
 
-  deleteTask(id)
+    deleteTask(id)
+  }
 }
 
-async function getTasks() {
-  return await fetch('http://localhost:3000/api/lists/1/tasks?all=true')
-    .then((response) => response.json())
-    .then((tasks) => tasks.forEach((task) => appendTaskToList(task)))
-    .catch((error) => {
-      console.error(error)
-      response.status(404).json(error)
-    })
+function getTasks() {
+  return fetch('http://localhost:3000/api/lists/1/tasks?all=true').then(
+    (response) => response.json()
+  )
 }
 
-async function postTask(task) {
-  return await fetch('http://localhost:3000/api/lists/1/tasks', {
+function postTask(task) {
+  return fetch('http://localhost:3000/api/lists/1/tasks', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
     },
     body: JSON.stringify(task),
   })
-    .then((response) => response.json())
-    .catch((error) => {
-      console.error(error)
-      response.status(404).json(error)
-    })
 }
 
-async function updateTask(id, done) {
-  return await fetch(`http://localhost:3000/api/lists/1/tasks/${id}`, {
+function updateTask(id, done) {
+  return fetch(`http://localhost:3000/api/lists/1/tasks/${id}`, {
     method: 'PATCH',
     headers: {
       'Content-Type': 'application/json',
@@ -176,8 +148,8 @@ async function updateTask(id, done) {
   })
 }
 
-async function deleteTask(id) {
-  return await fetch(`http://localhost:3000/api/lists/1/tasks/${id}`, {
+function deleteTask(id) {
+  return fetch(`http://localhost:3000/api/lists/1/tasks/${id}`, {
     method: 'DELETE',
     headers: {
       'Content-Type': 'application/json',
